requirements.txt
discord.py
sqlalchemy
alembic
asyncpg
bidict
emoji
cachetools
-e .